Add continious sampling

+++ src/utils/utils.py
@@ -336,7 +336,7 @@ def hms_collate_fn(batch):
         elif isinstance(v[0], np.ndarray):
             v = np.stack(v)
             if k == 'label':
-                v = v.astype(int)
+                v = v.astype(np.float32)
             v = default_collate(v)
             output[k] = v
         else:
